structural zeros added to the p1 model
several changes in several methods, all pertaining to structural zeros.
tested on directed, undirected, and mixed graphs. zeros are optional.
input as a graph. couple of developer notes left to consider: are we
going to pass this all the way down to bipartite.walk, the most
primitive function? probably should - see comments and decide.